refactor(extractors): rename law_dnc.py -> law.py
The `_dnc` suffix (divide-and-conquer) dates back to when both
`law.py` (legacy, 410 LOC) and `law_dnc.py` (current implementation)
existed in the tree.  The legacy `law.py` was deleted in Stream B9;
the `_dnc` suffix has been legacy noise ever since.

Scope
- `git mv src/refex/extractors/law_dnc.py src/refex/extractors/law.py`
  (git tracks as a rename; history preserved).
- Updated imports in `src/refex/extractor.py`,
  `src/refex/engines/regex.py`, `tests/test_law_extractor.py`,
  `tests/test_law_legacy.py`, `tests/test_edge_cases.py`.
- `tests/test_edge_cases.py`: renamed 10 `test_law_dnc_*` functions
  to `test_law_*` (pytest discovery still catches them); doc
  comments `# --- law_dnc.py:` -> `# --- law.py:`.
- Updated live-prose references in
  `docs/refactor2026/implementation_plan.md` (H3 note, §7 open
  follow-ups — removed the now-completed rename row) and
  `docs/refactor2026/optimization_log.md` (TODO list).
- Updated `CLAUDE.md` project-layout bullet.
- One-line reformat on the renamed file (ruff format).

Not touched
- Historical snapshot sections ("Ground Truth as of 2026-04-18"
  table, `CHANGELOG.md`, other `docs/refactor2026/*.md` planning
  docs) — those record state at an earlier point in time.

315 tests pass; lint + format clean.
