+++ src/refex/extractor.py
@@ -3,7 +3,7 @@ import re
 import warnings
 
 from refex.extractors.case import CaseRefExtractorMixin
-from refex.extractors.law_dnc import DivideAndConquerLawRefExtractorMixin
+from refex.extractors.law import DivideAndConquerLawRefExtractorMixin
 from refex.models import RefMarker
 from refex.orchestrator import CitationExtractor
